Integrated createUser mutation with database
