+++ server.ts
@@ -1,5 +1,14 @@
 import { ApolloServer } from 'apollo-server';
 import 'reflect-metadata';
+import { User } from './src/entity/User';
+import { AppDataSource } from './src/data-source';
+
+//Initializes database connection
+AppDataSource.initialize()
+  .then(async () => console.log('Database connection estabilished'))
+  .catch((error) => console.log(error));
+
+const userRepo = AppDataSource.getRepository(User);
 
 const typeDefs = `
 
@@ -50,12 +59,27 @@ const resolvers = {
     hello: () => `Hello World!`,
   },
   Mutation: {
-    createUser(_: unknown, args: CreateUserInput): UserOutput {
+    async createUser(_: unknown, args: CreateUserInput): Promise<UserOutput> {
+      const user = new User();
+
+      console.log('Inserting a new user into the database...');
+      user.name = args.userData.name;
+      user.email = args.userData.email;
+      user.password = args.userData.password;
+      user.birthDate = args.userData.birthDate;
+      await userRepo.save(user);
+      console.log(user.name, user.id, user.email, user.password, user.birthDate);
+      console.log('Saved a new user with id: ' + user.id);
+
+      //console.log('Loading users from the database...');
+      //const users = await userRepo.find();
+      //console.log('Loaded users: ', users);
+
       return {
-        id: 1,
-        name: args.userData.name,
-        email: args.userData.email,
-        birthDate: args.userData.birthDate,
+        id: user.id,
+        name: user.name,
+        email: user.email,
+        birthDate: user.birthDate,
       };
     },
   },
