+++ src/entity/User.ts
@@ -6,11 +6,14 @@ export class User {
   id: number;
 
   @Column()
-  firstName: string;
+  name: string;
 
   @Column()
-  lastName: string;
+  email: string;
 
   @Column()
-  age: number;
+  password: string;
+
+  @Column()
+  birthDate: string;
 }
