--- src/index.ts
@@ -1,22 +1,0 @@
-import { AppDataSource } from './data-source';
-import { User } from './entity/User';
-
-const userRepo = AppDataSource.getRepository(User);
-
-AppDataSource.initialize()
-  .then(async () => {
-    console.log('Inserting a new user into the database...');
-    const user = new User();
-    user.firstName = 'Timber';
-    user.lastName = 'Saw';
-    user.age = 25;
-    await userRepo.save(user);
-    console.log('Saved a new user with id: ' + user.id);
-
-    console.log('Loading users from the database...');
-    const users = await userRepo.find();
-    console.log('Loaded users: ', users);
-
-    console.log('Here you can setup and run express / fastify / any other framework.');
-  })
-  .catch((error) => console.log(error));
